Moved project to Applegate Standard Directory structure.

+++ image/stop-services.sh
@@ -58,9 +58,11 @@ export PASSWORD_JSON=$(cat password.json);
 
 echo -e "${white}Docker host is ${green}${DOCKER_IP}${nocolor}"
 
+pushd ../env/$1
 docker-compose stop
 docker-compose rm -f
+popd
 
-rm $(pwd)/traefik/traefik.toml
-
-docker ps
+if [ -f $(pwd)/traefik/traefik.toml ]; then
+    rm $(pwd)/traefik/traefik.toml
+fi
